Create Dockerfile to isolate project dependencies
The Dockerfile is meant to ultimately isolate the entire project. This
commit only cares about isolating dependencies related to the OCR
pipeline.

The requirements.txt was updated to include the necessary python
packages. The .dockerignore file was created to avoid sending large or
sensitive files and directories to the docker daemon and image.