Run OCR pipeline on docker container
Moved the OCR_pipeline notebook to the repository and now it can be run
using the project's container. Also added a section in the notebook
about spell checking.

Created a docker-compose file. The idea is to use it to build and run
our project, however I didn't manage to make it work completely.

Small changes to Dockerfile and requirements.txt