Set up container development environment
Dockerfile is based on a straight ubuntu 18.04 image now. Also tried to
use a straight alpine 3.11 image but didn't manage to do it (code is
left commented).

Updated OCR_pipeline.ipynb to assess whether the pre-trained model for
Jamspell exist so it doesn't cause the ipython kernel to die.

devcontainer.json is crucial for the development environment. It sets
up the vscode extensions, the mounting of the project directory and the
Jupyter notebook server.