Fine-tune sBERT model and some fixes
The FineTuneEmbeddingModel notebook does the fine-tuning of a sBERT
model using the existing labeled sentences on the task of predicting
the label of each sentence given its embedding. As the embeddings
contribute for the loss functions, gradient descent ends up updating
them. A linear classifier is learned to map the sentence embeddings to
the output label.

The LatentEmbeddings_ZeroShotLearning notebook had some issues fixed
(e.g. A PCA with 50 PCs retained will have the same 1st and 2nd PCs as
a PCA with 5 PCs retained). Besides, a section where the new fine-tuned
model is evaluated was added.