Create Mississippi spider - unfinished
The page components aren't in HTML. A curl request had to be done and
I suspect some javascript components will need to be dealt with in
order to obtain the pdf from each query result.

Right now the main problem is that we don't get a link to each pdf
document. Also I wasn't able to run the spider because the custom
modules weren't being imported correctly.