Fine-tuning Notebook with **Early Stopping** feature
This notebook has the following features

1. Overriden model.fit() and model._eval_during_training() functions to stop the training after the current epoch's accuracy is lower than 5 previous accuracies OR the difference between current epoch's accuracy and previous accuracy is less than 0.01. These paramters can be set in the notebook in the cell where these functions are overriden.

2. The model is saved in the Google Drive and can be loaded later on. In google drive, the structure for doing the experiments is

	Model_results
		|
		- BinaryClassificationExperiments
			|
			- 1
				|
				- Saved Model
				- 1_FineTuningResults.json
				- Confusion matrix and other saved figures
			- 2
			- 3

Right now there is only Experiment 1 but as others start fine tuning and experimenting it would be organized to conform to one folder only.
For MulticlassClassification you can create a new directory by changing the code in this notebook where the directory is defined.

**Only 1 model is saved per experiment folder**.

The aim of this is to make the fine tuning process more efficient and organized for the whole team.