Update mexico.py
There was a litle bug. There were a couple of variables that were made available through "self" to different functions that were used sequentially.
One of these variables was a date. The date was set un the fisrt function and later used in another function. By that time the date had already changed. The solution was to pass the values of these variables throug the kwarg of the yield function.